Fixed paths and moved app init to app folder
