--- stock_dashboard_api/__init__.py
@@ -1,9 +1,0 @@
-from flask import Flask
-from stock_dashboard_api.config import DevelopmentConfig
-
-app = Flask(__name__)
-
-app.config.from_object(DevelopmentConfig)
-
-
-from stock_dashboard_api import routes
