+++ stock_dashboard_api/app.py
@@ -1,4 +1,5 @@
-from stock_dashboard_api import app
+from app import app
+
 
 if __name__ == '__main__':
     app.run(debug=True)
