+++ stock_dashboard_api/app/config.py
@@ -2,7 +2,7 @@ import os
 from dotenv import load_dotenv
 
 project_folder = os.path.expanduser('~/StockDashboard/StockDashboardAPI')  # adjust as appropriate
-load_dotenv(os.path.join(project_folder, '.env'))
+load_dotenv(os.path.join(project_folder, '../.env'))
 
 
 class Config:
